+++ source/arra.h
@@ -1,6 +1,6 @@
 #ifndef ARRA_H
 #define ARRA_H
 
-double arra(int m, int n) noexcept(false);
+long long arra(int m, int n) noexcept(false);
 
 #endif
