+++ source/comb.h
@@ -1,6 +1,6 @@
 #ifndef COMB_H
 #define COMB_H
 
-double comb(int m, int n) noexcept(false);
+long long comb(int m, int n) noexcept(false);
 
 #endif
